requirements.txt
torch==2.4.1
torchvision==0.15.0
numpy==1.23.5
matplotlib==3.7.1
Pillow==9.4.0

